update: 更新文件 avrecorder_napi.cpp

+++ frameworks/js/avrecorder/avrecorder_napi.cpp
@@ -1787,7 +1787,8 @@ int32_t AVRecorderNapi::CheckRepeatOperation(const std::string &opt)
 
 int32_t AVRecorderNapi::GetAudioAacProfile(const int32_t &mime, AacProfile &aacProfile)
 {
-    if (mime >= AacProfile::AAC_LC && mime <= AacProfile::AAC_H2_V2) {
+    if (mime >= static_cast<int32_t>(AacProfile::AAC_LC) &&
+        mime <= static_cast<int32_t>(AacProfile::AAC_H2_V2)) {
         aacProfile = static_cast<AacProfile>(mime);
     } else {
         aacProfile = AacProfile::AAC_LC;
